Made Bedogni2024 params variable, introduced proper validation tests.
This took an embarrasing time: I managed to break the SCF code by
mis-typing the Ohno parameter (added a square)...

It was probably a good thing to carefully revisit the Bedogni comparison
now, and it checks the Huckel and SCF result, and also has a little play
with varying the parameters, and seeing that the calculation still
works.

This is all ground work so we can start to ML the params.