Rewrite model interface in universal 2nd quant form
Now each model file has a unified t and gamma for constructing the PPP
Hamiltonian in the 2nd quantisation form. Essentially this is the
physics-style notation of Zhang 2011, for the very motivated intent to
include this model (Carbon only) next, and then start thinking about
optimising model parameters for a particular system.

Doing it this way simplified the PPP code considerably. And, in the end,
did not need any edits to the CI code (as it should).

Atoms object is now overcomplicated, as is the read_geometry function.
But when I tried to strip it out I ended up with bogus results, so
clearly there is some weird dependency there.

Anyhoo, it is late and I am tired, so enough for one sesh!